CombinedRateProviderFactory
Adds contract for CombinedRateProviderFactorty and CombinedRateProvider. The purpose of these contracts is to combined two rates in scenarios where the common denomination of assets is not uniform. For example, a rate provider denominated in wstETH and another in wETH cannot be used in a pool pairing. Combining the wstETH denominated rate with wstETH's known rate provider solves this issue with a new contract.